refactor: tidy table cache and tests
