+++ tablecache.go
@@ -1,27 +1,19 @@
 package rindb
 
-//lint:file-ignore U1000 The table cache will be wired into the DB in a later step.
-
 import (
-	"container/list"
 	"context"
 	"errors"
 	"fmt"
-	"sync"
 	"sync/atomic"
 	"time"
-
-	"golang.org/x/sync/singleflight"
 )
 
-// --- You already have this type in your project ---
-// type SStable struct {
-//  *FileSystem
-//  SparseIndex SparseIndex
-//  Bloom       *BloomFilter
-// }
+const (
+	defaultShardItemCapacity    = 256
+	defaultCorruptMapCapacity   = 16
+	defaultTombstoneMapCapacity = 16
+)
 
-// --- typed errors ---
 var (
 	ErrClosed     = errors.New("tablecache: closed")
 	ErrObsolete   = errors.New("tablecache: obsolete")
@@ -61,6 +53,8 @@ type tableCacheOptions struct {
 }
 
 // FDLimiter is a minimal semaphore interface (Acquire before Open, Release after Close).
+// FDLimiter is a minimal semaphore interface (Acquire before Open, Release after Close).
+// TODO: provide concrete implementations.
 type FDLimiter interface {
 	Acquire(ctx context.Context) error
 	Release()
@@ -99,61 +93,6 @@ func (h *Handle) Unref() {
 
 // --- SLRU internals ---
 
-type segment uint8
-
-const (
-	segNone segment = iota
-	segProbation
-	segProtected
-)
-
-type entry struct {
-	key    tableKey
-	h      *Handle
-	seg    segment
-	elem   *list.Element // element in prob/prot list
-	pinned bool          // pin top-N / critical tables
-}
-
-type shard struct {
-	mu    sync.Mutex
-	items map[tableKey]*entry
-
-	// SLRU segments: probation & protected
-	prob list.List
-	prot list.List
-
-	// per-segment soft caps (by bytes)
-	probCapBytes int64
-	protCapBytes int64
-
-	// per-segment tracked usage (actual bytes)
-	probBytes int64
-	protBytes int64
-
-	// tracked usage (actual bytes) of entries in lists/map
-	usedBytes int64
-	capBytes  int64
-
-	// admission stop — set during Close() to reject new installs
-	stopAdmission atomic.Bool
-
-	// corrupt quarantine: key -> expiry
-	corrupt map[tableKey]time.Time
-
-	// tombstones to prevent re-admission after Delete/obsolete
-	tombstone map[tableKey]struct{}
-
-	// per-shard singleflight for open de-dup
-	flight singleflight.Group
-
-	// backref to parent (for global byte accounting)
-	parent *tableCache
-
-	// metrics (local)
-	hits, misses, opens, closes, evicts, promotions atomic.Int64
-}
-
 type tableCache struct {
 	shards    []*shard
 	shardMask uint64 // len(shards)-1, power of two
@@ -181,9 +120,9 @@ func newTableCache(opt tableCacheOptions) *tableCache {
 	}
 	for i := range shards {
 		s := &shard{
-			items:        make(map[tableKey]*entry, 256),
-			corrupt:      make(map[tableKey]time.Time, 16),
-			tombstone:    make(map[tableKey]struct{}, 16),
+			items:        make(map[tableKey]*entry, defaultShardItemCapacity),
+			corrupt:      make(map[tableKey]time.Time, defaultCorruptMapCapacity),
+			tombstone:    make(map[tableKey]struct{}, defaultTombstoneMapCapacity),
 			capBytes:     per,
 			parent:       c,
 			probCapBytes: int64(float64(per) * fr),
@@ -196,8 +135,10 @@ func newTableCache(opt tableCacheOptions) *tableCache {
 	return c
 }
 
+// shardFor deterministically assigns a table key to a shard using a simple
+// 64-bit mix of the DB and file numbers. The number of shards is always a
+// power of two, so a bitmask is sufficient to pick the shard without a modulo.
 func (c *tableCache) shardFor(k tableKey) *shard {
-	// Cheap 64-bit mix; shards is power-of-two so we can mask.
 	h := (k.DBID * 11400714819323198485) ^ (k.FileNum * 14029467366897019727)
 	return c.shards[h&c.shardMask]
 }
@@ -482,158 +423,6 @@ func (c *tableCache) Close(ctx context.Context, drainTimeout time.Duration) erro
 		}
 	}
 	return nil
-}
-
-// --- SLRU helpers (all require s.mu held unless stated) ---
-
-func (s *shard) promoteOnHit(e *entry) {
-	switch e.seg {
-	case segProbation:
-		// second hit → promote to protected
-		if e.elem != nil {
-			s.prob.Remove(e.elem)
-		}
-		s.probBytes -= e.h.actualBytes
-		e.elem = s.prot.PushFront(e)
-		e.seg = segProtected
-		s.protBytes += e.h.actualBytes
-		s.promotions.Add(1)
-
-		// enforce protected cap via demotion if necessary
-		for s.segmentBytes(segProtected) > s.protCapBytes {
-			if tail := s.prot.Back(); tail != nil {
-				dem := tail.Value.(*entry)
-				if dem.elem != nil {
-					s.prot.Remove(dem.elem)
-				}
-				s.protBytes -= dem.h.actualBytes
-				dem.elem = s.prob.PushFront(dem)
-				dem.seg = segProbation
-				s.probBytes += dem.h.actualBytes
-			} else {
-				break
-			}
-		}
-	case segProtected:
-		// move to MRU
-		if e.elem != nil {
-			s.prot.MoveToFront(e.elem)
-		}
-	}
-}
-
-func (s *shard) unlink(e *entry) {
-	switch e.seg {
-	case segProbation:
-		if e.elem != nil {
-			s.prob.Remove(e.elem)
-		}
-		s.probBytes -= e.h.actualBytes
-	case segProtected:
-		if e.elem != nil {
-			s.prot.Remove(e.elem)
-		}
-		s.protBytes -= e.h.actualBytes
-	}
-	e.elem = nil
-	e.seg = segNone
-}
-
-func (s *shard) chooseVictim() *entry {
-	// Prefer tail of probation; fall back to tail of protected.
-	if back := s.prob.Back(); back != nil {
-		return back.Value.(*entry)
-	}
-	if back := s.prot.Back(); back != nil {
-		return back.Value.(*entry)
-	}
-	return nil
-}
-
-// segmentBytes returns the tracked byte usage for the requested segment.
-func (s *shard) segmentBytes(seg segment) int64 {
-	switch seg {
-	case segProbation:
-		return s.probBytes
-	case segProtected:
-		return s.protBytes
-	default:
-		return 0
-	}
-}
-
-// evictOrDemoteLocked ensures segment splits and capBytes.
-// It closes unpinned, zero-ref victims outside the lock to avoid blocking.
-func (s *shard) evictOrDemoteLocked() {
-	if s.capBytes <= 0 {
-		return
-	}
-	// First, if protected exceeds its cap, demote from protected tail into probation head.
-	for s.segmentBytes(segProtected) > s.protCapBytes {
-		if tail := s.prot.Back(); tail != nil {
-			dem := tail.Value.(*entry)
-			if dem.elem != nil {
-				s.prot.Remove(dem.elem)
-			}
-			s.protBytes -= dem.h.actualBytes
-			dem.elem = s.prob.PushFront(dem)
-			dem.seg = segProbation
-			s.probBytes += dem.h.actualBytes
-		} else {
-			break
-		}
-	}
-
-	// Then evict until within total cap.
-	var toClose []*Handle
-	for s.usedBytes > s.capBytes {
-		v := s.chooseVictim()
-		if v == nil {
-			break // nothing to evict; budget won't be met but we're out of candidates
-		}
-		if v.pinned {
-			// skip pinned victims: keep them MRU-protected to avoid tight loops
-			if v.seg == segProbation && v.elem != nil {
-				s.prob.Remove(v.elem)
-				s.probBytes -= v.h.actualBytes
-				v.elem = s.prot.PushFront(v)
-				v.seg = segProtected
-				s.protBytes += v.h.actualBytes
-			} else if v.seg == segProtected && v.elem != nil {
-				s.prot.MoveToFront(v.elem)
-			}
-			continue
-		}
-		// Remove from SLRU + map; stop future pins; free budget immediately (cache residency accounting).
-		s.unlink(v)
-		delete(s.items, v.key)
-		s.usedBytes -= v.h.actualBytes
-		s.parent.totalBytes.Add(-v.h.actualBytes)
-
-		if v.h.refs.Load() > 0 {
-			// Busy: mark for close later when refs drain.
-			v.h.evictWhenZero.Store(true)
-		} else {
-			// Cold: close now (outside lock).
-			toClose = append(toClose, v.h)
-		}
-		s.evicts.Add(1)
-	}
-
-	if len(toClose) > 0 {
-		s.mu.Unlock()
-		for _, h := range toClose {
-			s.closeNow(h)
-		}
-		s.mu.Lock()
-	}
-}
-
-func (s *shard) closeNow(h *Handle) {
-	if h.closed.CompareAndSwap(false, true) {
-		_ = h.closer(h.Table)
-		s.closes.Add(1)
-	}
 }
 
 // --- Stats & utils ---
